// frontend/e-report/types/section-1/sectionOneSchema.ts
import { z } from "zod";

export const sectionOneSchema = z.object({
  branchCaseNumber: z.string().min(1),
  policeStationId: z.string().min(1),
  policeCaseNumber: z.string().min(1),
  status: z.enum(["DRAFT", "PENDING", "REJECTED"]),
});

export type SectionOneValues = z.infer<typeof sectionOneSchema>;